requirements.txt
pyserial>=3.4
requests>=2.25.1
python-dotenv>=1.1.1

pyarrow

# pandas
# pyarrow
# geopandas
# shapely